Refactor binding
