+++ Cargo.toml
@@ -12,6 +12,7 @@ repository = "https://github.com/nix-community/nix-user-chroot"
 [dependencies]
 nix = "0.26.2"
 libc = "0.2.139"
+clap = { version = "4.2.4", features = ["derive"] }
 
 [dev-dependencies]
 rstest = "0.17.0"
